+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+- Model class setting generator incorrectly leading to stop iterations. 
+
 ## [0.2.0] - 2018-08-21
 ### Added
 - Added the ability to pass custom arguments to the tqdm callback