+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Removed
 ### Fixed
 - Model class setting generator incorrectly leading to stop iterations. 
+- with_generators now has arguments ordered in a fashion consistent with for_steps (generators first, then steps).
 
 ## [0.2.0] - 2018-08-21
 ### Added